+++ Sources/Logger/Logger.swift
@@ -1,4 +1,3 @@
-import Either
 import Foundation
 #if os(Linux)
 import Glibc
@@ -103,21 +102,5 @@ public final class Logger {
     case warn
     case error
     case fatal
-  }
-}
-
-public func logError<A>(
-  subject: String,
-  logger: Logger,
-  file: StaticString = #file,
-  line: UInt = #line
-  ) -> (Error) -> EitherIO<Error, A> {
-
-  return { error in
-    var errorDump = ""
-    dump(error, to: &errorDump)
-    logger.log(.error, errorDump, file: file, line: line)
-
-    return throwE(error)
   }
 }
